+++ foundation/ringbuffer.h
@@ -16,20 +16,20 @@
 #include <foundation/types.h>
 
 
-FOUNDATION_API ringbuffer_t*   ringbuffer_allocate( unsigned int size );
+FOUNDATION_API ringbuffer_t*   ringbuffer_allocate( int64_t size );
 FOUNDATION_API void            ringbuffer_deallocate( ringbuffer_t* buffer );
 
-FOUNDATION_API void            ringbuffer_initialize( ringbuffer_t* buffer, unsigned int size );
+FOUNDATION_API void            ringbuffer_initialize( ringbuffer_t* buffer, int64_t size );
 FOUNDATION_API void            ringbuffer_finalize( ringbuffer_t* buffer );
 
-FOUNDATION_API unsigned int    ringbuffer_size( ringbuffer_t* buffer );
+FOUNDATION_API int64_t         ringbuffer_size( ringbuffer_t* buffer );
 FOUNDATION_API void            ringbuffer_reset( ringbuffer_t* buffer );
 
-FOUNDATION_API unsigned int    ringbuffer_read( ringbuffer_t* buffer, void* dest, unsigned int num );
-FOUNDATION_API unsigned int    ringbuffer_write( ringbuffer_t* buffer, const void* source, unsigned int num );
+FOUNDATION_API int64_t         ringbuffer_read( ringbuffer_t* buffer, void* dest, int64_t num );
+FOUNDATION_API int64_t         ringbuffer_write( ringbuffer_t* buffer, const void* source, int64_t num );
 
-FOUNDATION_API uint64_t        ringbuffer_total_read( ringbuffer_t* buffer );
-FOUNDATION_API uint64_t        ringbuffer_total_written( ringbuffer_t* buffer );
+FOUNDATION_API int64_t         ringbuffer_total_read( ringbuffer_t* buffer );
+FOUNDATION_API int64_t         ringbuffer_total_written( ringbuffer_t* buffer );
 
-FOUNDATION_API stream_t*       ringbuffer_stream_allocate( unsigned int buffer_size, uint64_t total_size );
-FOUNDATION_API void            ringbuffer_stream_initialize( stream_ringbuffer_t* stream, unsigned int buffer_size, uint64_t total_size );
+FOUNDATION_API stream_t*       ringbuffer_stream_allocate( int64_t buffer_size, int64_t total_size );
+FOUNDATION_API void            ringbuffer_stream_initialize( stream_ringbuffer_t* stream, int64_t buffer_size, int64_t total_size );
